Cambiar credenciales
Alta etapa

Incluir datos de prueba de vacs, enfermedades, alta de plan y etapa con su test

Terminar Test

Arreglar detalle en test

Agregar espacio para hacer legible

Agregar datos de etapa y plan en el load.sql

Arreglar long por int

Sacar fetch join al 2p

Test de crear agenda exitoso

Agregando consulta para filtrar etapas para un ciudadano

Usar parallelStream

Arreglar datos para las pruebas con nuevo modelo de restricciones

Funciona consulta

Hacer consulta mas eficiente

+++ sist-central/sist-central-ejb/src/main/java/datos/dtos/CiudadanoDTO.java
@@ -3,23 +3,23 @@ package datos.dtos;
 import java.io.Serializable;
 
 public class CiudadanoDTO implements Serializable{
-    private long ci;
+    private int ci;
     private String nombre;
     private String email;
 
     public CiudadanoDTO() {}
 
-    public CiudadanoDTO(long ci, String nombre, String email) {
+    public CiudadanoDTO(int ci, String nombre, String email) {
         this.ci = ci;
         this.nombre = nombre;
         this.email = email;
     }
 
-    public long getCi() {
+    public int getCi() {
         return ci;
     }
 
-    public void setCi(long ci) {
+    public void setCi(int ci) {
         this.ci = ci;
     }
